Component with private state transported to HOC
